Off-campus UCC Amendment (passed on 2/17/13)
Justification:
The mental health of all members of Ruddock House is incredibly important. In
the past off-campus members of Ruddock House have been encouraged to join an
on-campus alley. That alley’s UCC would become their default counselor.
However, on-campus UCC’s have often been unable to cultivate strong
relationships with off-campus alley members.

To increase the effectiveness of the mental health safety net, particularly for
off-campus members of Ruddock House, Off-Campus UCCs were introduced last year.
Off-Campus UCCs function as a traditional a on-campus UCC with one exception:
they must live off-campus. However, due to the lack of time it was impractical
to amend the Constitution to include Off-Campus UCCs.  Off-Campus UCCs have
been working for a term, and we believe their position should officially be
placed into the Ruddock Constitution. A clause has been added in the case where
no suitable volunteers can be found living off-campus; in this case, we allow
the Off-Campus UCCs to live on campus.

Also, due to the vague wording of “Elected Officer,” we have clarified the
meaning to be a member of Ruddock Excomm to allow Revcom Reps, Boc Reps, Stud
Bud, etc. to run.