ExComm (GovComm) Amendment (passed on 4/20/13)
Justification:

The current Ruddock house Executive Committee (ExComm) is composed of seven
votes divided across 13 people, with the offices of President, VP, Secretary,
Treasurer, and Librarian receiving one vote each, and the five-person Social
Team and three-person Athletic Team with one vote split across each team.

We believe that this ExComm structure seeks to accurately represent the
diversity of opinion and the class diversity we value in Ruddock House.
However, through lengthy conversation we believe we have identified specific
ways to improve the effectiveness of the ExComm:

Firstly we seek to add two positions: an Off-Campus and a Sophomore
representative. The Off-Campus representative will gather the views and
opinions of Off-Campus House Members and communicate them to the ExComm.
The Sophomore representative will help represent the Sophomore class, thus
allowing the librarian to focus more on representing the freshman class during
the latter part of his term.

Secondly we seek to consolidate the Social and Athletic Teams’ ExComm
representation with the introduction of the Social Chairman and Athletic
Manager positions, each to be held by a single person. This allows members of
Social Team and Athletic Team who are not ExComm members to pursue a position
as a UCC. It also allows the Social Chairman and Athletic Manager to be chosen
explicitly for their leadership abilities in their respective teams as well as
their leadership abilities on ExComm.

As a result, this amendment brings us to a 9-member, 9-vote ExComm, with
members of President, VP, Secretary, Treasurer, Librarian, Social Chairman,
Athletic Chairman, Sophomore Representative, and Off-Campus Representative,
each with a full vote.