VP/UCC Amendment (Passed on 12/12/13)
Justification:
As the duties of the Vice President overlap with those of an Upper Class
Committeman, it seems logical that an individual could be a UCC and then
run for Vice President. However, due to the difference in appointment
times of the two positions, this is currently a logistical nightmare.

The duties of the Vice President overlap with that of a UCC such that
the extra responsibilities would not prevent the successful execution of
both positions. Without this amendment, a new UCC may need to be
appointed for a single academic term, a process that is both difficult
and time-consuming leaving members of an alley without an official UCC
for that time.