Sophomore House Amendment (Passed on 12/12/13)
Justification:
In past years, there has been a greater number of sophomores wishing to
reside within the house than the number of beds available. This is
particularly true as sophomores are lowest in the pick order. To help
promote the continued involvement of the sophomore class, it would be
beneficial to have a residence explicitly designed for off-campus
sophomores. Additionally, we believe that designating sophomores’
priority for an off-campus location will help ease the transition
between on and off-campus and help foster a more vibrant off-campus
community.