Off Campus Rep Amendment (Passed on 12/12/13)
Justification:

The GovComm amendment created a new position called the Off-Campus
Representative. Currently, this officer is elected during the 2nd term
election cycle and serves during the following full academic year,
during which time he must live off campus.

This position is designed to bring the concerns and opinions of
off-campus members to the attention of ExComm and create more unity
between on-campus and off-campus Ruddock. It was also to be held by a
sophomore or junior to create balance in the class-distribution of
ExComm.

We believe that there are improvements to be made to the Off-Campus
Representative position. First of all, this amendment aligns the term
of the Off-Campus Representative with the term of the other elected
positions. Secondly, he would not be required to live off campus but
has to have lived off for at least one term.

This first change is important for several procedural reasons. As it
currently stands, any member elected to the position of Off-Campus
Representative would not be able to run for any other position in the
following year without vacating their current office. This amendment
would also avoid a complication of having two different compositions
of ExComm between the spring term of one year and the fall term of
the following year as this position turns over. In addition, the
previous Off-Campus Representative will still be attending Caltech
for the first half of the new Off-Campus Representative’s term, a
benefit that could not exist in the current system if a junior were
elected.

The second change is more philosophical. We believe the Off-Campus
Representative should not have to reside off-campus as long as
they already have in the past. This will encourage more members to
run for a position that doesn’t forbid them from living in
Ruddock.

It is essential for the Off-Campus Representative to understand
the difference between residing in Ruddock and residing
off-campus; however, anyone who has experienced this for at least
a single term will have a sufficient understanding. Therefore, we
believe anyone who has lived off-campus in the past, even if they
now reside in the house or will during part of their term, will
have the skills and understanding to hold this position.