Sophomore Room Pick Amendment (Passed On 04/10/14)
Justification:
As the position of Alluminati is now defunct, two room picks that
sophomores have traditionally held no longer exist. This is problematic
because within the Ruddock House room pick system, it is likely that
those picks would no longer be held by sophomores, greatly changing
the representation of each class within the House.

It is important that the House retain its relative balance of classes in
the interest of a healthy, diverse body. This amendment would give the
two room picks previously held by the Alluminati oﬃcers to two members
of the rising sophomore class selected by random number generation.