Soc Team Amendment (Passed on 12/11/14)
Proposed by: Elliot Simon
Endorsed by: Patrick Nikong, Galen Gao

Justification:

As currently stated in the constitution, the Social Team of Ruddock is in charge of, among
other things, organizing parties. Among these parties would be included Big Interhouse,
Our Private Interhouse, and Frosh Party. While this is the system used in many other
Houses at Caltech, Ruddock House has, in recent years, chosen volunteers not necessarily
on the Social Team to organize house parties.

It has been expressed both by members of the current Social Team, the current Executive
committee, and various other members of the house that system is not the method the
house desires. The office of Social Team has been elected without consideration of the
potential party organizers. Additionally, the Social Team turns over in the middle of the
construction of OPI which could easily cause paperwork consistency issues.

This amendment removes the language explicitly assigning the responsibility of organizing
house parties and dances to the Social Team. The Social Team shall still be responsible
for organizing all social functions. The process for choosing those individuals responsible
for organizing house parties shall be left unmentioned in the Constitution and as such shall
be chosen in a method deemed fit by the membership of the House.