Social Member Qualifications Amendment (Passed on 12/11/14)
Proposed By: Elliot Simon
Endorsed By: Patrick Nikong, Allika Walvekar

Justification:

Currently, in order to become a Social Member of Ruddock, a candidate must satisfy the
following conditions: They must be an undergraduate at Caltech, be voted on by the
procedures described in the Constitution, and live neither in Ruddock nor in a designated
off-campus alley. This third condition was originally designed as an extra clause preventing
non-full members of Ruddock House from living on campus or in an off campus location
designated for Ruddock members. In other words, a student desiring to live in Ruddock
House would have to apply for full membership, not social membership.

This amendment would simply remove the restriction on the current residence location from
the requirements of social membership. Thus a social member must be an undergraduate
at the school and have been elected as such as detailed below. This will allow an
upperclassman who currently resides in the Ruddock House Off Campus Alley to become a
social member and thus participate in house activities.