Amendment Procedures Amendment (Passed on 12/11/14)
Proposed By: Elliot Simon
Endorsed By: Allika Walvekar, Galen Gao

Justification:

This amendment makes two substantial changes to the amendment procedure. The first is
that in addition to being posted immediately following the House Meeting at which the
amendment is presented, the amendment must be sent out electronically 3 days before the
House Meeting. The amendment is sent out so that all House Members have a chance to
read the substance and explanation of the amendment before the discussion takes place at
the House Meeting.

The second change adds a 2nd option for obtaining the endorsement of the amendment. In
addition to having the amendment endorsed by a 1/5 of the Full Membership, an
amendment can also be proposed with the endorsement by 2 members of the Executive
Committee. This number was chosen to match the number of members required to call a
meeting of the Executive Committee as in 3.5.1. This allows a much faster path while still
requiring a check on the amendment approval. The old endorsement pathway is kept as an
option to put a check on the power of the Executive Committee. Should the committee not
want to endorse an amendment due to incompetency or political motivation, amendments
can still be proposed by the entire house.