Picks Order Amendment (Passed on 2/24/16)
Proposed By: Elliot Simon
Endorsed By: Tim Liu, Henry Swanson

Justification:

In the spring of 2013, an amendment was passed that fundamentally changed
the composition of the Executive Committee (Excomm) of Ruddock. Excomm
used to be composed of the top five, ath team, and social team where the
athletic team and soc team had votes proportional to their team size. This
created a '7 vote' Excomm composed of 13 people. The new Excomm
replaced Soc Team with the newly created Social Chairman, Ath Team with
the newly created Athletic Manager, and added sophomore representative and
off-campus representative. However, the picks order and description
remained untouched.

This amendment adjusts the picks order so that it is realigned with what
is most likely the original intent. The picks order goes Excomm, UCCs,
then interhouse positions, then minor house positions by importance. As
such, the Ath Team and Social Team are moved down to below the interhouse
positions while Athletic Manager and Social Chairmen are left in their
place. This is done by creating a separate section for vote order from the
section describing the guaranteed picks.

Note in the guaranteed picks section, the Social Team and Social Chairman
combined are guaranteed 3 picks among the entire group. They made divide
these picks however they choose with the default ordering being that as
detailed in section 7.1.4 of the constitution. As such, the Social
Chairman and Athletic Manager are not guaranteed one of the picks of the
Social Team and Ahtletic Team respectively. This is because members of
excomm can live off campus (like OMG the off campus rep) and as such do
not need to be guaranteed a pick. If you disagree, Elliot would like to
hear why.