Soc/Ath Team Election Amendment (Passed on 2/24/16)
Proposed By: Elliot Simon
Endorsed By: Tim Liu, Henry Swanson

Justification:

As everyone at elections this year remembers, the Ath Team and Soc Team
votes combine took nearly half of the total election meeting time. While
this was exaggerated due to a couple extenuating circumstances, these two
elections still take a disproportionately large part of the election
meeting.

The solution to this problem implemented below is to hold the Ath Team and
Soc Team elections separate from the rest of the election cycle. These
would be held exclusively via absentee ballot run by elect comm starting
after the conclusion of the election meeting and ending 72 hours later.

By separating out the elections, the election meeting will be much
quicker. By conducting the Ath Team and Soc Team elections via absentee
ballot, each individual may take as long as they choose to create their
preference for some or all candidates without needing to write down such
votes and place them in a hat. In addition, the elections can easily be
counted or simulated via a program with little difficulty from electcomm
on a separate date. The main constraint on running such an election is the
time it takes to run the full instant-runoff, not the complexity. As such,
this system is lower stress and less time consuming for everyone involved.

Note that the period for grievances against the results of any election
would commence at the conclusion of the election for Soc Team and Ath Team
and run for 36 hours after that. While this means that there is a delay
between the conclusion of the election of most positions and the
resolution of grievances, it avoids having to create 2 such periods.