Election Signup Deadline Amendment (Passed on 2/24/16)
Proposed By: Elliot Simon
Endorsed By: Tim Liu, Henry Swanson

Justification:

This amendment places a deadline on when you can sign up for an elected
position. This deadline is set as 36 hours prior to the start of the
elections meeting. Should the timing for ath and soc teams be changed,
these teams would still need to be decided 36 hours before the start of
the normal election meeting.

The goal of this is to create an absolute time at which point the list of
candidates is completely set before elections begin. This enables Absentee
voters to place their votes without any possibility of the ballot changing
after their vote was submitted. Since many people have real and reasonable
conflicts with the election meeting and cannot attend, it is reasonable to
create an opportunity for them to express their full preference on all
candidates.