Secretary and Treasurer On-Campus Amendment (Passed on 5/7/16)
Proposed By: Nikita Sirohi
Endorsed By: Tim Liu, William Schmidt

Justification:

Ruddock in the past used to have a culture where living on campus was seen
as largely necessary for someone to extensively participate in house
activity. Over the past six years or so, this has been changing. The
Treasurer and Secretary are generally influential, active, enthusiastic
house members.

If a Treasurer or Secretary would prefer to live off-campus, allowing them
to do so would be an excellent way to continue the push Ruddock has been
making to better integrate off-campus Rudds.

The job of the Secretary is to take minutes, send out emails, interface
with the housing office, and work with alumni. As a former Secretary, I
personally didnÎéÎ÷t feel that any of these duties required me to live
on-campus. The job of the Treasurer is to serve as the house steward and
to take care of house funds. While one can argue the job of steward is
made easier by living in the house, this amendment does not remove that
option from the Treasurer, merely gives the Treasurer the ability to
decide whether or not it is necessary for them. Out of the four Treasurers
in recent memory, two have been elected while living off-campus, and it
does not seem to have prevented them from doing an excellent job.