Condorcet Voting Amendment (Passed on 12/13/16)
Proposed By: Henry Swanson
Endorsed By: Karim Lakhani, Tim Liu

Justification:

Last year, there were a few interesting incidents involving voting systems behaving in unexpected ways. The Review Committee had a problem with determining the winners of the BoC Secretary election, because it was not clear how to fairly determine a two-winner election. Also, in the Ricketts Secretary election, there were six candidates, and the final ballot moved a candidate (the Condorcet winner, in fact) from first to fourth. This inspired some discussion about our voting methods, and several amendments were passed last year. I intended to write this amendment around that time, but then never did, because I was really busy and/or lazy.

The goal of this amendment is to move away from instant-runoff voting. It may be better than first-past- the-post, but it is one of the most fragile and misbehaved ranked voting methods (see the examples at the bottom of this document). There are many better ranked voting methods out there, but my personal favorite, both in terms of fairness of results and ease of computation, is the Ranked Pairs method. It always elects a Condorcet winner if one exists, encourages honest voting*, and can easily be done on a whiteboard. An explanation of the method can be found at Wikipedia or Electorama.

Lastly, this amendment consolidates the BoC Representative voting section
with the rest of the Procedural Amendment. The method for the BoC reps is
still fundamentally based on instant-runoff, because I couldn't find any
good hybrid of Condorcet and proportional-representation methods. Again,
come to the meeting if you've got any ideas.

* It's impossible to completely avoid strategic voting, as per the Gibbard-Satterthwaite theorem, but to manipulate this method, the amount you'd need to know about how other people are going to vote is too much to be practical.